# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BR Tours and Travels API
  version: 0.3.0
  description: >-
    Express business API for the public website, authenticated administration,
    media, enquiries and notification operations. Booking requests are leads,
    not payments, issued inventory or guaranteed reservations.
servers:
  - url: /api/v1
...
paths:
...
  /admin/media:
    get:
      tags: [Admin content]
      security: [{ cookieAuth: [] }]
      responses: { "200": { description: Bounded protected media library } }
    post:
      tags: [Admin content]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfToken" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, altText]
              properties:
                file: { type: string, format: binary }
                altText: { type: string, maxLength: 300 }
                visibility: { type: string, enum: [PUBLIC, PRIVATE] }
      responses:
        "201":
          {
            description: "Valid raster decoded, metadata stripped and re-encoded to WebP, or a detected PDF stored for attachment-only delivery",
          }
        "415": { $ref: "#/components/responses/Error" }
  /admin/media/{id}:
    get:
      {
        tags: [Admin content],
        security: [{ cookieAuth: [] }],
        parameters: [{ $ref: "#/components/parameters/Id" }],
        responses:
          {
            "200": { description: Protected media metadata record },
            "404": { $ref: "#/components/responses/Error" },
          },
      }
    patch:
      {
        tags: [Admin content],
        security: [{ cookieAuth: [] }],
        parameters:
          [
            { $ref: "#/components/parameters/Id" },
            { $ref: "#/components/parameters/CsrfToken" },
          ],
        responses:
          { "200": { description: Media metadata or visibility updated } },
      }
    delete:
      {
        tags: [Admin content],
        security: [{ cookieAuth: [] }],
        parameters:
          [
            { $ref: "#/components/parameters/Id" },
            { $ref: "#/components/parameters/CsrfToken" },
          ],
        responses:
          {
            "204": { description: Unreferenced media removed },
            "409": { $ref: "#/components/responses/Error" },
          },
      }
  /admin/media/{id}/file:
    get:
      {
        tags: [Admin content],
        security: [{ cookieAuth: [] }],
        parameters: [{ $ref: "#/components/parameters/Id" }],
        responses:
          {
            "200":
              {
                description: Protected private or public media bytes; PDFs use attachment disposition and sandbox CSP,
              },
          },
      }
  /media/{id}:
    get:
      servers: [{ url: / }]
      tags: [Public content]
      summary: Public media bytes
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          {
            description: Explicitly public raster bytes or attachment-only PDF bytes,
          }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: br_admin_session
      description: Environment-configured opaque session cookie; never a browser-storage token.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, maxLength: 30 }
...
    CsrfToken:
      in: header
      name: x-csrf-token
      required: true
      schema: { type: string }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              {
                type: object,
                additionalProperties: { type: array, items: { type: string } },
              }
            requestId: { type: string }
  responses:
    Error:
      description: Safe API error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }